pyproject.toml
[tool.poetry]
name = "utilities"
version = "v1.0.0"
description = ""
authors = ["Vikash Pathak <vikash@logiciel.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "1.13.2"
annotated-types = "0.7.0"
anyio = "4.4.0"
async-timeout = "4.0.3"
asyncpg = "0.29.0"
click = "8.1.7"
exceptiongroup = "1.2.2"
fastapi = "0.112.0"
h11 = "0.14.0"
idna = "3.7"
pip = "24.2"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
python-dotenv = "1.0.1"
setuptools = "70.3.0"
sniffio = "1.3.1"
starlette = "0.37.2"
typing-extensions = "4.12.2"
uvicorn = "0.30.6"
wheel = "0.43.0"
psycopg2-binary = "^2.9.9"
aws-sam-cli = "^1.123.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
